Probe and save oVirt API path
This patch changes the oVirt provider so that it will determine the path
of the oVirt API probing the candidates. Once determined it will be
saved to the "path" column of the "endpoints" table, and the next time
it is required it will be retrieved from the table without probing.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>


(transferred from ManageIQ/manageiq@15c5c03d1261c872210a5f37fa11070cecaf8991)