Do not add the /ovirt-engine prefix to oVirt ems_ref
Currently when version 4 of oVirt is added as a provider, the "ems_ref"
attribute of the objects stored in the database contains the "/ovirt-engine"
prefix, because that is what the "href" attribute returned by oVirt contains.
As a result, depending on the version of oVirt this "ems_ref" attribute will be
different: /ovirt-engine/api/* for version 4 and just /api/* for versoin 3.
This complicates migration from version 3 to version 4 of oVirt, as the both
prefixes need to be supported. To avoid that problem this patch changes the
provider so that the "ems_ref" will be calculated always in the same way,
regardless of the version of oVirt used. To keep backwards compatibility with
objects already stored in the database the value will be /api/*, as it used to
be with version 3 of oVirt.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>


(transferred from ManageIQ/manageiq@7ed0cabcfaf6e5540d92fb9869e6792b80c32f8a)