Select `ovirt` or `ovirt-engine-sdk` gems with `version` parameter
This patch changes the `connect` method of the `Redhat` provider so it decides
which gem to use to connect to the oVirt engine based on the new `version`
parameter. If the version parameter is given and the value is 4, then the
`ovirt-engine-sdk` gem will be used, otherwise the `ovirt` gem will be used.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>


(transferred from ManageIQ/manageiq@fec6a166477581c798a1beefb1547d496022531b)