Omit the version part from the OS name
The convention of other providers and the expected format
of the OS field name is to contain the name only, without the
version.

The version is being concatenated to the OS name on the UI
side, therefore the code which did it on the provider side is
removed.

http://bugzilla.redhat.com/1360416


(transferred from ManageIQ/manageiq@8141e663cec228821b270a3101ddbff3fbe54e67)