Adjust memory reconfiguration for oVirt
Changing the memory assigned to virtual machine in oVirt must follow the
following rules:

1. If the virtual machine is running, then the difference btween the old
amount of memory and the new amount of memory needs to be a multiple of
256 MiB. The reason for this is that memory is added plugging new
virtual DIMMS to the virtual machine, and the size of these is 256 MiB.

2. The physical memory reserved for the virtual machine should always be
less or equal to the virtual memory.

This patch changes the oVirt provider so that implements these rules,
adjusting the amount requested by the user accordingly. For example, if
the virtual machine is running with 100 MiB of virtual memory, and the
user requests to increase it to 300 MiB, then the provider will adjust
this to 356 MiB, so the delta will be a multiple of 256 MiB.

https://bugzilla.redhat.com/1356468
https://bugzilla.redhat.com/1356475


(transferred from ManageIQ/manageiq@6bbdab40e66e1b56211cb48d171f8271b6a0758d)