Update cluster when modified
When we change cluster for a specific vm we were not able to
fetch proper cluster information. We were not able to get this
information from the event due to code limitation.

Now we fetch all the clusters and datacenters so we have enough
information to update the vm.

This patch fixes:
https://bugzilla.redhat.com/1397503


(transferred from ManageIQ/manageiq@80f3bf987354f80740bb1e290b02a36e35b5c1e2)