rhv: removed the option to migrate the VMs outside of the cluster.
The support for cross cluster migrations was added to oVirt only as a
workaround for el6->el7 migrations but should not be exposed. Since the current
code in manageiq anyway did not work properly, removing the support for it
completely - it is a low level functionality, it is obsoleted and discouraged
to be used.

Links
https://bugzilla.redhat.com/show_bug.cgi?id=1398287


(transferred from ManageIQ/manageiq@2fbf505a904bf56bb707958ee717e32c2f7310cd)