Save host for a Vm after migration
Fix saving the host of a Vm after it was migrated when doing
targeted refresh of the Vm.

Before if the hosts inventory was not fetched sperately by the
refresh, the host of a VM was not updated properly even though
its uid was returned as part of the target Vm inventory.

Now we take the information from the Vm inventory and use it
to update the Vms host properly.

https://bugzilla.redhat.com/show_bug.cgi?id=1368925


(transferred from ManageIQ/manageiq@e346b46655f672c64458266ebe5ccb36f075c056)