Don't pass empty lists of certificates to the oVirt SDK
The oVirt SDK accepts lists of trusted CA certificates. But when an
empty list, or a list containing just 'nil' elements is passed, the
result is that no certificate is trusted, not even the system default
trusted certificate. To avoid that this patch changes the oVirt provider
so that it passes 'nil' instead of empty lists.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>


(transferred from ManageIQ/manageiq@118b7ec52fb9336d7b7af59115da6442fd4db1a2)