Move code from manageiq-gems-pending
Now that the oVirt provider has been extracted to a separate repository,
it makes sense to move all the oVirt specific code from
'manageiq-gems-pending' to this repository.

This patch moves the 'ovirt_event_monitor.rb' and 'ovirt_inventory.rb'
files. These files implement event monitoring and inventory collection,
using version 3 of the oVirt API and the 'ovirt' gem. As there is an
ongoing effort to use version 4 of the oVirt API, and the oVirt SDK,
these files have been added to a new 'ManageIQ::Providers::Ovirt::Legacy'
module, to make it explicit that this code will eventually be removed.

https://github.com/ManageIQ/manageiq-providers-ovirt/issues/8

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>