Fix fetching ISO images from non active domains
We used to ignore the data-center level when looking for an ISO domain,
as a result we would sometimes take an ISO domain which was not attached to
an active data-center.
We now first find a data-center which is "up" and then take an ISO domain which is
attached to it and is active.
This fix is applied for providers that support version 4 of the api.

https://bugzilla.redhat.com/show_bug.cgi?id=1447560