Select datastore by its association with the provider
Datastores on the same setup may have the same name.
However, there is an assumption that datastores names on the same
provider are unique. The selection of datastore by its name should be
narrowed to the provider on which the vm is provisioned and has at least
one host with write access.

https://bugzilla.redhat.com/show_bug.cgi?id=1427498