Replace 'get_provider_destination' with 'with...'
Currently the 'get_provider_destination' method runs a block, using the
'with_provider_object' method, and from within that block it returns the
object that was yielded to the block. That object is associated to a
connection, and the connection is closed when the block finishes. The
net result is that the next time that the 'get_provider_destination'
method is called, for the same object, it returns an object associated
to a closed connection, which then fails. To avoid that issue this patch
replaces the 'get...' method with an equivalent 'with...' method that
does not return the object.

This patch is part of the fix for the following bug:

  Cloud-init not working when using api v4
  https://bugzilla.redhat.com/1486340