Add connection manager
This patch adds a new 'ConnectionManager' singleton intended to simplify
re-use of connections in single threaded processes. This class will
manage a set of connections, indexed by the identifier of the EMS. The
connections will be created when needed, re-created when options change,
and closed when the worker finishes or when the EMS is removed from the
database.

Checking for removed EMSs is performed periodically, with an interval
controlled by the new 'connection_manager.purge_interval' parameter. The
default value is 1 hour:

  :ems:
    :ems_ovirt:
      :connection_manager:
        :purge_interval: 1.hour

The existing code is also changed so that it uses the connection manager
instead of creating the connections directly, and so that it does not
close the connections explicitly.

This only applies to V4, the V3 code does not change.