Assume 3 when can't determine API version
Currently when using version 4 of the SDK is enabled the provider raises
an exception when it can't determine what versions of the API are
supported by the oVirt server. But this is a normal situation. It
happens in general when the credentials of the server are not known yet,
for example when the oVirt server has been discovered scanning a range
of IP addresses. If an exception is reaised then the discovery process
will fail. To avoid this issue this patch changes the provider so that
it tolerates this situation, and assumes that the server supports at
least version 3 of the API.

This patch addresses the following bug:

  RHV41 Provider Discovery failure
  https://bugzilla.redhat.com/1502963

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>