Use db query to infer supported api_versions
The probing for api versions was done by connecting to the provider directly
which causes problems when done from a UI applience which has no access to the
provider.

We now use the "api_version" (which actually contains the product version)
to infer the api version supported. When the feild is not available we try probing.