Running 'ensure_network provider' from oVirt's refresh
'ensure_network' has an api call to oVirt, therefore, it should be run
as a separate jobs to avoid calling it directly from the ui.

oVirt's infra provider refresh is running in a worker. The creation/update
ofthe network manager was moved to this flow.

The tests were fixed accordingly.