Try both API versions in `raw_connect`
Currently the `raw_connect` method needs to receive the version of the
API in order to work correctly. If it doesn't receive it it fails
because the method that determines it just doesn't exist (it is an
instance method, but `raw_connect` is a class method). To avoid that
this patch changes the `raw_connect` method so that it tries with both
versions of the API.

This patch fixes partially the following bug:

  Failed validation when adding RHV provider
  https://bugzilla.redhat.com/1509432