Reload provider when api_version available
Before, when an instance of the ext_management_system was memoised before api_version is set would
probe the provider again and again to get list of available versions.

Now it checks if api_version is available in db and reloads it if so.
This fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1512490