Handle correctly custom cloud-init script
The change that implemented cloud-init with version 4 of the API
introduced an error in the handling of custom cloud-init scripts: all
the cloud-init attributes are transformed into Ruby symbols, thus they
have a leading `:` character that cloud-init does not understand. For
example, if the following cloud-init customization template is used:

  #cloud-config
  write_files:
    - path: /tmp/test.txt
        content: |
        Here is a line.
        Another line is here.
      permissions: '0755'

It will be translated into the following *incorrect* cloud-init custom
script:

  :write_files:
  - :path: "/tmp/test.txt"
    :content: |
      Here is a line.
      Another line is here.
    :permissions: '0755'

To fix that this patch changes the `update_cloud_init!` method so that
it does the conversion from strings to symbols only for the attributes
that need to be removed from the script and treated specially.

This change addresses the following bug:

  [REGRESSION] Cloud-init provision on RHEV
  https://bugzilla.redhat.com/1518341

Related-To: https://bugzilla.redhat.com/1486340
Related-To: https://github.com/ManageIQ/manageiq-providers-ovirt/pull/88
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>