Implement template targeted refresh
Since we can create template from MIQ side on RHV, templates should be
fully refreshed once the creation is completed.
In addition, when a template is being created on RHV side, not all of
its fields are being updated.
Therefore, there is a need to support also template target refresh
(unrelated to graph refresh in this context).

Fixes bug:
https://bugzilla.redhat.com/show_bug.cgi?id=1517852