Update to version 4.2.0 of the oVirt Ruby SDK
This version of the SDK is backwards compatible with the previous one,
but it also introduces support for the features that have been added to
oVirt in version 4.2.

The more relevant change in the SDK itself is the addition of exceptions
specific for authentication failures and timeouts, which are convenient
to simplify error handling in the provider.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>