Fix Seal option of publish VM
Due to latest rename of property name in publish vm yaml file from :seal
to :seal_template, sealing option wasn't processed properly.

In addition, the property should be send as part of the request header,
therefore should be specified in the second parameter of the
Template.add(template, {:seal => ...}) method.

https://bugzilla.redhat.com/show_bug.cgi?id=1514895