Fix location of `pipeline` and `connections` settings
The `pipeline` and `connections` settings are in the following path:

  :ems_refresh:
    :rhevm:
      :pipeline: 40
      :connections: 10

But the part of the code that creates the SDK connection gets them from
a different place. As a result the parameters arent really applied, and
the number of connections is unlimited. When combined with the settings
used by the asynchronous refresh code, this results in the creation of
hundreds of connections. When there are more than 256 connections open
the oVirt server will start rejecting them, and the refresh fails.

This patch fixes the location of the parameters in the code that creates
the SDK connection.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>