vm.hardware.nics[i].lan should return the network attached to the vm
The 'lan' table contains a mapping of the the networks attached to the
hosts.

If a vm is running, vm.hardware.nics[i].lan should return the host network
it is attached to.

There are two reasons it wasn't working -

1. host_inventory didn't store the same instance of the lan in the 'hash'
and in 'lan_uids'.
Since the vm_inventory used the 'lan_uids' instance, it didn't get the
id of the lan once it was stored to the db (the id was updated on the
'hash' instance).

2. The vnic (vm nic) returned by the ovirt sdk doesn't contain the network
name (the value is always null). The network should be retrieved using the
vnic profile.

* Similar fixes were done to the relevant files in graph refresh.
* v3 wasn't fixed.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1572917