Fix broken network spec
The ovn spec was broken due to openstack provider changes.
Made a new recording.

Note currently the ovn did not respond to some of the requests that had
limit=1000 as a parameter, this is a bug on the ovn side and will be
handled separately, the current recording was actually modified as if
the bug did not exist. This does not reflect on the changes introduced
in this commit but I thought it is important to note.