Add support for CA for Ansible role
When a CA string is passed as an extra_var create a temporary directory
and in it create a file with the string as its content.
Pass the path of the file as the engine_cafile to the Ansible role.

This is part of implementing: https://bugzilla.redhat.com/show_bug.cgi?id=1644605