Fix iscsi storage duplication
Storages are identified by their location attribute, which in the case
of the iscsi was based on the id of the first lun in the volume group.
Order of the returned luns is not guarnteed so this would sometimes
change if the luns were changed on the provider side.

We will now base the location on the volume_group id.