Memory and CPU reconfigure applied at the same time
RHV engine decides which changes can be applied live leaving the other
for the 'next_run'.
If a next_run configuration is already pending, the new update request
must be sent with the 'next_run' option set to true.
This is necessary to merge new specs with all the others waiting to be
merged.

resolve #416