Native viewer support
Add native viewer support to allow the download of the connection file.
In case of multiple available consoles the "spice" one is consider the
preferred and if spice is missing a message is logged.
In case of headless vm (with no consoles available) an exception is
raised.

(cherry picked from commit 6e127f2fc77ab98f4d9772070d3573d58c4d0406)