Fix for datacenters being recreated every refresh
Datacenters subclass EmsFolders and thus can't be two separate
collections since they are the same table.

If you had for example two folders and a datacenter the
persister.ems_folders collection would have the two folder inventory
objects but the query manager.ems_folders would return three records
including the datacenter.

This meant that the datacenter was being deleted by the ems_folders
inventory collection, then the datacenters inventory collection was
creating it.

This fix is to save all records for the table in the base collection and
manually set the STI :type column (which was actually already being
done).